Add files via upload
Ideation notes, paer protype and playtesting feedback. Since the RetroTwistGameJamFullDetails document said an optional submission would be opened, our team was under the impression that we would have to submit it to Canvas. However, since no assignment has been created on Canvas so far, we are now uploading it here instead.